Update apt cache regardless
On brand new instances the last modified date of the apt-cache file is
the age of the instance. That kinda negates the point of looking up a
last-modified timestamp. Also fixed an issue with zsh. The install
scripts are run with bash regardless, so it doesn't matter whether zsh
is the parent executor.