Overload `gf` mapping for Node module resolution
Typically `gf` works in files by looking at the filename under the
cursor and attempting to resolve it relative to the cwd (`:pwd`), but
that doesn't work well in JavaScript. Or at all.

Now if you open a JavaScript file it'll overload `gf` with a new
function, checking the contents of the string and using the same file
path resolution node uses at runtime.