Move other files into `autoload/`
Files are executed in order at startup via `linked/init.vim` and an
ordered set of autoload calls. Some no-op functions were added to each
file so it had something harmless to call.

That's probably the least clean part of it. The rest has been a massive
success! Yay for the plugin-oriented vimrc!