Explicitly set neovim clipboard provider
I can't live like this! On occasion the `pbcopy` command comes in handy,
but 95% of the time it's just a nuisance. Thankfully, like most things
in vim, it's configurable.

Also, turns out `pbcopy` and `pbpaste` were the culprits for horrible
lag I've experienced on mac, and always just assumed it was a mac thing.
It's such a relief to see faster speeds. I feel like I'm editing in
Ubuntu again!