Clock/timezone fixes & rando improvements
Set the tmux built-in clock to use 12h mode, found a way to expose the
`$TZ` variable to the whole tmux process, and removed it from the
time-printing statusbar utility thing.

Also added a `:Details` vim command to show who wrote the given line,
when, and how they tried to explain themselves (almost always
incriminating).