Add eject script
I made this tool for working in a vm, but it's been so useful I wanna
make it part of the framework. It copies all the files into a
timestamped directory and tars it up.

Super great for things you don't wanna commit, but still need to keep
around in automated backups or when you're about to revert a vm.