Update readme
Make it clearer that the purpose of this project is to make my own life
easier. People are 100% free to cannibalize the source, but I won't
support or consider them when writing oodles of breaking changes daily.

Also be less informative about installation. Nobody should ever install
it with the exclusion of me but I already know this because I share a
psychic connection with myself.