Migrate to zsh for associative arrays
Bash 4 has a neat feature: hash tables. It pairs beautifully with my
eject script. Unfortunately mac doesn't support it (GPL v3 reasons), and
I don't feel comfortable hacking it in. Too many things depend on it.
I'm also not enthused about doing shebang hackery to pull from homebrew
or native between platforms.

Thankfully zsh supports hash tables. They're a tiny bit different, but
if the time comes to migrate, it should be a 1-line change (ha, like
that ever happens).

Instead of a list of files to package, now it takes a map of
`destination_name => source_path`. Previously the destination name was
implied, but made the process almost irreversible. These changes allow
for a restore script which unpackages a tarfile and resettles all the
files.