Add support for ignored commits in blames
Now `git#blame#()` respects a special variable:
`git#blame#ignored_commits`. It's a list of commit hashes to ignore. Why
is this useful? Well...

- Code formatting changes (e.g. `prettier` or `rustfmt`)
- Codemods (ES2015, module path renames, sweeping API changes)
- Indentation changes (usually happens with code formatters)

If one of the git blames is an ignored commit, `git#blame` queues up
a query to dig deeper in the commit history. This operation is recursive.

I've been wanting this option pretty much since the first day I built
`:Author`.