Disable ALE in node_modules
Because `further.vim` is so fancy, I often end up editing some file in
`node_modules` (not because I'm insane, just as a debugging technique).
It gets annoying because ALE tries to format and lint everything and
won't stop complaining.

This uses a little-known feature of ALE to disable it given a file path
pattern.