Add better symlink manager
Rust doesn't fit well in this framework. At least, not in it's present
form. Making changes and adding features/errors was just prohibitively
hard.

This (finally) replaces the Rust symlink process with a simpler bash
script. Soon the Rust code will be gone, and a new `dotfiles-env`
symlink script will be supported.