Add lookup logic for test => source file
Now I can navigate the other way! If I start in a test file, now I can
jump to the source file. It gets a little crazy. First, it checks the
grandparent directory for something with the same name. If that doesn't
work (unfortunately that's the case for many projects I work in), then
it parses the test file looking for relative imports that resemble the
test file's name.

I'm pretty excited to try this!