Add test command
Now that I can infer the test file, it's a rather short leap to
automatically running the tests in another pane. That's something
I already do in one of my private `dotfiles-env` presets, but it was so
tailored to my workflow that it's practically useless on general
projects.

Still, every time I'd move out of my VM I'd reach for `:Test` and
grimace when it wasn't there.

This is the more polished form. General gist: running `:Test` will
locate the test file, create a tmux pane on the right if it doesn't
already exist, then starts a new test process in watch mode focused
solely on the relevant test file.