Replace `fzf` vim plugin with `skim`
Go full Rust. I'm currently getting a non-deterministic error with fzf.
I think it's related to this kernal bug:
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1813873

It seems to happen more often with `fzf` than `skim`, but that could
just be in my head. Replacing it with `skim` just to be sure, and also
go full Rust.