Switch to antigen
Replaces a ton of install logic with a few lines of `antigen`.
I recently learned about that plugin manager. It's pretty nice.

Oh, and I also deleted all the lua installation code. Not because
I needed to or because it was a good idea, rather I felt like it and
this is my world so deal with it.