Remove vader test scripts
These tests are waaaaay too brittle. I've written a reasonable amount
and can safely say tests in vimscript are nowhere's near as valuable as
they are in other languages. There's almost no mocking, stack traces are
almost non-existent, the headless environment is a pain and a half to
configure on CI, and only a tiny subset of commands can actually be
tested. The rest require E2E style tests. It's too much work for such
a little payoff.