Add nix env vars to fish shell
My zsh config was able to source a file in the nix profile which
configured the path for all the install nix binaries, but that didn't
work for fish. It's not posix compliant.

There's probably a better way. I keep seeing reference to `nix-shell`
and it sounds closer to what I really want, but that's a challenge for
another day.

The end result of that nix script is adding a directory to your path.
Easy enough to replicate.