Fix linux install for fnm and upgrade to latest
The tarball isn't a tarball, it's a zip file. It worked fine on mac
(which is where I wrote this script) but failed miserably on linux
because it couldn't unpack the zip file.

I suspect it has something to do with BSD's `tar` command and the
built-in `unzip` support. Purely guessing.

Anyway, I've replaced the tar download with a generic URL and I'm
manually unzipping with `unzip` instead of `tar`. It seems to work
great. The real question is whether it works on mac. TBD.