Hack: load vim plugins through nix
I've added a bridge between `vim-plug` and a nix derivation that allow
me to load all my plugins from the nix store. That's kind of awesome!

TBH I'm not too proud of it though. It's a hack, and I'd like cleaner
ways of doing it. For one, vim plugins already exist in the nixpkgs
repo. Any plugin manager I build should work with what's already there,
and it should speak the common language: derivations.

I did this mainly for compatibility with the vim-plug methodology.
Theoretically I could automatically wrap the built-in plugins with
a bash script that would extract the plugin path into `$out`, or maybe
even symlink to the real store path (which should work with their
dependency detection).

Anyway, there are *much* better ways of doing it, but IT WORKS!