Move vim settings into init.vim
I'm in the process of completely renovating how I install vim plugins.
The way it was set up before was an abomination that never got cleaned
up. See, a long time ago I moved my vimrc to a plugin, but it still
needed immediate hooks that ran before other plugins, so after jamming
it in `&rtp` I forced a `runtime` check on specific files.

That was obviously a bad idea.

This commit moves several things from `settings.vim` into the `init.vim`
file with the goal of eliminating the `&rtp`/`runtime` hacks. Not
everything was copied over. I audited all my settings, dropped some, and
tweaked others. It's mostly the same though.