Replace antigen plugin manager with Nix
This isn't as fancy as the vim plugin system. As far as I can tell,
there's no way to inject "plugins" into zsh. All plugin managers seem to
do is download and source a zsh file of the same name. If Nix handles
the download, there's no reason to have a plugin manager hang around
just for the ceremony of sourcing a file.

All plugins moved into `pkgs/zsh-plugins`.