Integrate with host clipboard outside neovim
I've written a tiny utility for managing neovim's clipboard integration.
It uses a "virtual provider" whenever tmux is active and a "host
provider" in other cases.

The core difference between this and my old vimrc is that now it uses
the host provider (`pbcopy`/`pbpaste`) when those are defined and tmux
is unavailable.

A notable difference is the assigned clipboard integration is repeatable
through a function, allowing me to dynamically switch clipboard
providers. I anticipate using this feature to replace the `<leader>t`
copy mode toggle.