Turn editor/ into a plugin bundle
You know what's better than one vim plugin? Hundreds! Er, not quite. But
historically I build up hacky plugins in the dotfiles framework as
a proof of concept and later pull them into their own repository. That's
how alternaut and debrief started.

It would be nice to have another layer of separation between all the vim
plugins actually demarcating boundaries.

This commit changes the structure of `editor/` so that every child of
the directory is treated as an isolated vim plugin. It even has it's own
generated vim package.