Add jq dependency
I've been using it for a while, but it never made it into my standard
tool set. I took the time to read the man page and was amazed by how
useful it was. It solves most of what I'd hoped.

The only thing I'm missing now is a data conversion tool, something that
can take json, xml, msgpack, etc and convert it to other formats. `rq`
looks promising, but the project appears unreliable at the moment.