Extract desktop environment config to NixOS module
This was one of the bigger challenges with using my dotfiles module on
other computers. I might want the editor and some networking tools, but
use it primarily as a headless server. Now that the desktop is extracted
to a module, I can easily disable it all together and pick the toolkits
piecemeal.

Partly addresses https://github.com/PsychoLlama/dotfiles/issues/8