Write vim plugin for managing Zettelkaesten
I've been having trouble managing the sheer quantity of notes. I gave up
a complicated system a while ago and went back to plain markdown notes
and directories, but it has become a nightmare to manage the
hierarchies. The Zettelkasten method appeals to me, so I'm testing it
out.

Also I'm not putting this in the vim plugin nursery because with the
advent of Flakes, I don't think it makes sense anymore. I just haven't
gotten around to reintegrating them.