Revamp SwayWM keybindings
I've dropped a lot of keybindings that came with the default config as
I never ended up using them. I've also tweaked a lot of keybindings to
be more similar to macOS. I've had to use that platform as a work
computer and switching between the two is enough to cause a stroke.

Hopefully with an external keyboard I can make the experience similar
enough to work.