Tweak SwayWM output positions
I've had trouble every since c1fd2ba. After a reboot, sway refused to
use the external monitor and Rofi started rendering offscreen. I fixed
it by setting my built-in monitor to a fixed position of `0 0` and
positioning the external display relative to that on negative space,
meaning the output coordinates for my laptop never change, only external
displays.

Rofi continued being a pain. When the larger screen was plugged in, it
sized itself accordingly even though it was rendered on the smaller
screen. I managed to fix that by explicitly stating my built-in
monitor's dimensions.