Remove mcfly
It was a fun experiment and not a bad shell integration, but I just
don't use it enough to justify the shell cost. More notably, this seems
to interact poorly with ZFS on my dev machine, causing massive ARC
CPU/RAM spikes when I create a new shell.

I should figure out why, but I won't.