Upgrade to NixOS 22.05
I finally fixed the blockers! The main one was git. It wanted me to
explicitly say dotfiles was a trusted repository when executing from
root. I'm not thrilled about the implementation of that feature, but
I've got ideas on how to improve it in my framework.

The other problem was an OpenGL compatibility problem with the newer
version of Alacritty. Turns out I was upgrading the wrong channels and
the packaged version of Alacritty was depending on a newer GL feature
than NixOS shipped. Rookie mistake, but I didn't catch it until later.