Import git.vim from nursery
I'm rejoining the vim plugin nursery with my dotfiles framework. It was
a fun experiment but it didn't pan out. Now it's time to end it.

Note: I've removed support for recursively ignored commits. I'm using
a version of git that supports `--ignore-rev` now.

I've also removed dependence on my stack trace parser and assert
utilities. I'd like to keep this as trim as possible.

Someday I'll rewrite it in Lua...