Set global registry & path to flake inputs
I've been meaning to do this for a while. This sets the `pkgs` flake and
`<nixpkgs>` channel to match `inputs.nixpkgs-unstable`. Now I can test
something in the shell and easily add it to my dotfiles, certain that
I'm actually testing the real thing. This would've saved me some time
when troubleshooting a build error with `fzf` a few commits ago.