Add overlay for bleeding-edge packages
I'm migrating off of my bespoke modules onto home-manager's system.
I can write modules inside the `home-manager.users.[]` namespace, but it
doesn't provide `nixpkgs-unstable` as a module argument and for the sake
of portability, I'd rather not add it.

This takes a different approach: instead of overriding packages
individually with `package = unstable.my-package`, I'm applying it
globally through an overlay.

This might be a bad idea. I'm not sure yet.