Move most package overrides into presets
The latest package overlay experiment was a bad idea. It grew quite
difficult to keep things in sync with the modules that used them, and it
only seemed to be getting worse. Further, I'm not sure how I'd opt out
of the override short of another overlay, which is pretty heavy-handed.
I'd rather set this as `mkForce`.

This works around the special `nixpkgs-unstable` argument by sticking it
on `pkgs.unstable` instead, which is carried through the home-manager
module.