Branch modules on architecture
There are 3 platform targets:

- home-manager
- darwin
- nixos

They all depend on which modules are loaded. This decides what I can use
as building blocks, and why I'm migrating to the `home-manager`
framework (it has a very solid cross-platform foundation).

If you're on Darwin, now the framework should load the Darwin module,
and if you're on NixOS, well, it loads the NixOS module. No shock there.

Until I can split these out in a serious way it's still loading the same
module under the hood.