Use a more standard SwayWM config
This is adapted from a hack branch that's still WIP attempting to
migrate to home-manager's Sway module. They generate an `input` field
for me, and this aligns my own custom module with their implementation.
It also removes the custom session variables that aren't supported and
replaces them with standard Sway input configuration.