Abandon multiple editors approach
I originally configured `programs.editors` to allow multiple editors
simultaneously. That was a bad idea. I don't see a "clean" way of
identifying the primary editor, which is necessary for man pages and the
`nvim` binary. I also didn't have a good reason other than "it sounded
cool".

To be fair, it was pretty cool.