Upgrade to NixOS/HM 23.05
Now I'm running on NixOS 23.05. The only upgrade difficulty I had with
the operating system was that my ssh-agent service no longer accepted
passwords. Not great. I'm not sure exactly how I solved it, but
I created a new key, asked it to add that one in particular, and it
worked. After that, it happily accepted the password for my old key as
if nothing ever happened.

Before I got it working I booted into an older version and verified the
problem only existed in 23.05.

As part of the home-manager upgrade, I had to alter a good deal of
nushell config. They fixed the config location for Darwin which allowed
me to delete all my custom modules and rely on the standard options.

Another minor thing is that newer versions of nushell require constant
values for argument defaults. I was using an expression in the `p`
command.