Remove compiled editor from shell
This has caused me to lose so much time. Trying to debug why changes
aren't showing up only to realize the direnv cache was not invalidated
and didn't recompile the flake to create a slightly different version of
my editor... it does not spark joy.

I've got better ideas on how to approach this.