Lazily load all Neovim plugins
Now all vim plugins are loaded in a package as `opt/` instead of
`start/`. An initial lua script loads them immediately after the vimrc
to preserve behavior with a little compat layer in between.

The goal: be able to conditionally disable and replace vim plugins at
runtime. Right now it is impossible to disable a plugin. Once it's added
to `start/`, you cannot undo it.

This makes it very troublesome to edit my vim plugins. In the past I've
basically recompiled my editor without that plugin, added it to a Nix
shell, and linked the new plugin in the local environment. It's horrible
and has all kinds of downsides (e.g. overriding the editor with an older
version, sometimes changes not taking effect because direnv didn't
invalidate the Nix flake, slow setup time, difficult to combine multiple
in-development plugins...)

My next objective is to add support for `direnv.vim` or something
similar. A set of environment variables that conditionally disable and
replace vim plugins. This way I can keep the same editor with sensible
defaults and strict control, but break out whenever I need to develop
more interactively.