Make vim plugins dynamically configurable
By setting a special environment variable I am able to add or replace
vim plugins. This allows me to automatically use a local version of
a Vim plugin by changing into that directory, which makes development
infinitely easier. I used to have to compile a new version of my editor
that disabled the plugin and hard-coded the path into `&rtp` and that
carried a ton of downsides. It also slowed down my Nix shells
profoundly.