Tweak Nushell UX flags
A couple of changes: now all tables are displayed using the PostgreSQL
syntax. I like that display better. And now it shows header names on the
footer after 20 items (previously 25). I've also boosted the max history
size to 100 billion because history is culture, and culture must
survive.

Lastly, I changed the cursor to a line in insert mode because it looks
cool, idk. Maybe I'll set it back.