Remove hacky bits from dotfiles substrate
One of my original "bright ideas" was to alias `dotfiles.user` to
whatever the primary user account was, and that way my nixos/darwin
presets could modify the user, such as adding it to groups and what not.

Terrible idea. Thankfully I only used it in one place, and only to add
the user to a group. It only applied to one system anyway. I can easily
inline that.

There aren't really any commonalities between nixos and darwin in my
configuration. Anything that's shared should go in the flake or the
`define-host` lib.