Reuse pkgs set for all host builders
Instead of passing `system` and letting the host builder import nixpkgs,
now I'm requiring every host definition to pass pkgs directly. This
means I don't have to duplicate the overlay set. A common source of bugs
is forgetting to add an overlay in one of the imports and realizing
a not-found error days later when I go to build a different component.

This will pose a challenge for consumers of this flake. I have a couple
other flakes that build on this, and they won't have access to
`pkgsBySystem`. I'll figure it out.