Disable Wezterm ligatures
Seemed like a cool idea, but in practice it's very annoying. I don't
want the text moving around while I'm typing. Other features like
combining `==` and `!=` is disconcerting, because I'm no longer certain
what the line does without memorizing the ligature. Which is not a large
investment, but I don't see the payoff.