Restore nushell TS queries to working state
The NixOS upgrade brought a new version of the Nushell grammar that
broke my highlight and indent queries. The names of AST nodes changed
(for the better, IMO) but that meant my old queries were throwing an
error on every keystroke. Not great.

This fixes the queries enough to restore functionality.