Allow coc and native LSP to coexist
This only defines the LSP bindings that clobber `coc.nvim` if the
attached LSP is not Copilot, which is an LSP that attaches to every
buffer.

I can unconditionally define global mappings once coc is dead. But in
the meantime, this allows me to upgrade one piece at a time.