Update nushell config
I'm not quite sure why, but restarting the computer seems to have a more
permanent effect at restarting nushell. It's as though tmux is keeping
a daemon alive somewhere, and the config is only re-evaluated after
a reboot.

I don't know.

Whatever the case, this recovers the system by migrating the config to
a supported mode.