Conditionally load configs alongside plugins
This is one of the intended outcomes of migrating to a core framework
package loader: associating config files with plugins, so if
I conditionally disable something like `telescope-nvim` it won't still
try to load its config file.

The config is exposed as `plugin.config` in the loader so I can still
continue loading it if I want to swap implementations for another plugin
(e.g. a dynamic `&rtp` injection).