Restore support for local plugin overrides
This adds support for RTP-based plugins by declaring `'path'` as the
plugin type, which will load `plugin.source` instead of assuming it
lives in the packdir.

I'm happy with that part.

I've also restored the `VIM_PLUGINS` override behavior, which was always
a hack. I'm on not happy with that part.