Remove unused plugins
I haven't written Unison in a while. There's probably a treesitter
grammar for it now, or will be by the time I return.

Vader, I'm probably not going to need that again. I'm on board with Lua
now.

Vim Plug hasn't been used in ages. I'm amazed it was still in there.