Replace EFM server with nvim-lint
I'm not as impressed with nvim-lint as I am with conform, despite
conform praising nvim-lint as the inspiration. I may end up returning to
EFM.

The main reason to switch is observability. Language servers make it
more difficult to figure out why a command is failing, or if it's
failing at all. The hope is that nvim-lint makes it easier to
troubleshoot problems with ESLint and friends.