Disable completion for nvim cmdline
I like it, but it messes up three key things: First is `Man` completion,
which I use extensively. Second is paging through history. It seems to
get stuck. Third is VimL completion which is very weak.

I'll miss it, but it doesn't seem stable enough yet.

On the flipside was lua completion. It did phenomenal for `lua=`
expressions. I may restore it just for that pattern, if possible.