Ensure nvim standard lib runs first
I spent a bit trying to re-add frontmatter support to markdown files
only to discover that the TS injections were still in `$VIMRUNTIME`. It
seemed something was overriding them. I'm not sure which package is
responsible, but it hardly matters. Packages should be extending the
core `$VIMRUNTIME` not the other way around.

Making sure it's the first entry of `&rtp` fixes frontmatter injections,
along with many other features I'm sure I haven't noticed.