Disable tmux systemd service
The equivalent never worked quite right on macOS, which was the whole
point of it (permissions). Well, that and logs and service monitoring.
But I discovered that's why `wl-paste` and friends never worked in my
terminal. Clients inherit the server's environment, and if the systemd
service starts before the graphical session OR operates in an
independent environment, it will never receive those environment
variables.

By removing the tmux server I can finally run those commands again.