Disable Wayland for Wezterm
Wezterm was working until I disabled fractional scaling, then windows
were supposedly being created but never appearing. Logs looked fine.
I found this issue and discovered that it works if I open rofi and do
Super+Return to launch the terminal. Seems floating windows allow it to
work.

This is a workaround suggested in the thread but new builds should have
it resolved. I'll wait for a while and see if I can remove it.