Namespace all presets & profiles under GH username
This is a continuation of the previous commit. Now all custom presets
and profiles are namespaced as `psychollama.*`. This makes it easier to
import and selectively enable modules in other flakes, such as my home
lab, without conflicting on other repo's presets and profiles.

I do the same thing in my home lab. Everything's namespaced `lab.*`.