Fix nesting of presets and profiles
It seems when I refactored this to include my username on presets and
profiles I mixed up the ordering. Instead of "username/presets" I did
"presets/username". Thankfully I defined it correctly in the options,
but files should mirror the module system.

Now it's all consistent. The happy side effect is my exported
`nixosModules` no longer require multiple imports to pull in all the
configs. I remember in the beginning that adding `profiles` would create
an evaluation error because I forgot to add it in the flake. Now it's
all consistently defined where I need it.