Disable cursor blink mode
Apparently there are two ways the cursor blinks. Once via
`visible-cursor` which is always on (why). Twice via the
`blink-cursor-mode` but only if you start an action and cancel it,
like `SPC ESC`. Or something like that. I didn't really understand it
but I do know there is never a time ever when you should ever blink
the cursor ever for any reason at all. Ever.