Add direnv plugin for emacs
Now I'll have access to linters and formatters. This is a lot more
important when running in daemon mode because it won't inherit the
shell environment. It also fixes a long-standing issue I had with
neovim, where switching projects in the same editor instance would
throw errors because it wasn't started from that project directory.

Man, emacs is great.