Go back to Telescope instead of snacks.picker
I tried to like it. There's just so much that's missing. The most
annoying aspect is it replaces the wrong buffer. Open navitron, launch
the file picker, and it replaces the split pane on the other side
leaving navitron open.

The speed difference is barely noticeable. I'm going back to Telescope.

This reverts commit a9891698463bf16b383253d7fd9d4aa3c2dd3c26.